+++ egs/fluent_speech_commands/SLU/transducer/train.py
@@ -26,7 +26,7 @@ import torch
 import torch.multiprocessing as mp
 import torch.nn as nn
 import torch.optim as optim
-from asr_datamodule import SluDataModule
+from slu_datamodule import SluDataModule
 from lhotse.utils import fix_random_seed
 from torch import Tensor
 from torch.nn.parallel import DistributedDataParallel as DDP
